Config editor can now run as package
