+++ qudi/tools/config_editor/__main__.py
@@ -1,0 +1,28 @@
+# -*- coding: utf-8 -*-
+"""
+Qudi is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+Qudi is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with Qudi. If not, see <http://www.gnu.org/licenses/>.
+
+Copyright (c) the Qudi Developers. See the COPYRIGHT.txt file at the
+top-level directory of this distribution and at <https://github.com/Ulm-IQO/qudi/>
+"""
+
+import sys
+from .config_editor import ConfigurationEditor, ConfigurationEditorMainWindow
+
+app = ConfigurationEditor(sys.argv)
+# Init and open main window
+mw = ConfigurationEditorMainWindow()
+mw.show()
+# Start event loop
+sys.exit(app.exec_())
